kernel: fix bug in ext2fs driver
The code allocated a static amount of memory to store the ext2 raw
inode, can then copied in the amount specified by the file system.
This is fine when the on disk inodes are the same size as the defined
C structure, but it doesn't have to be in all ext2 file systems. In
fact, new versions of genext2fs don't, but the one of my local machine
did.