build: optimize various build parameters
This patch includes using a minimal build of the native tools when
building the os, which means that to get the native tools, users
should now have a separate build directory which is invoked using
normal cmake.

This patch also modifies the toolchain build scripts to run make with
as many jobs as there are CPUs by default.