toolchain: define USE_LD_AS_NEEDED in GCC configuration
This should make the toolchain be much better behaved. Before, it
would fail to link against libgcc_s even when the executable needed
called its __register_frame_info() function. This was "fixed" by linking
against libgcc_s in libc, but this created a new problem, becasue libgcc_s
also depends on libc. The construction of this cycle caused the dynamic
linker to enter an infinite loop. This issue should also be fixed, but
this toolchain patch removes a build dependency and is also used on other
systems.

This patch also fixes the CI automatic build of the toolchain, which previously
was missing the outputs field when checking the outputs.