kernel: start refactoring boot code to allow different boot protocols
kernel_main() now takes in a boot info type argument, which is for
now asserted to be BOOT_INFO_MULTIBOOT2. This patch also updates the
x86 boot code to support the Xen boot protocol (PVH) for qemu.

Qemu doesn't actually implement Xen, and so only looks at the
PHYS32_ENTRY tag, which it uses to load the kernel like a multiboot1
kernel (except with xen info structure instead of multiboot). The
motivation for supporting this is so that qemu can be passed the
-kernel option instead of requiring having a grub2 iso. Also, supporting
multiple boot protocols also would allow porting to the fancy new
bootloader limine.