kernel+tests: allow running automated tests through Qemu
This essentially works by replacing the normal /bin/start program
with a single test program, which is then executed directly be the
kernel. This program outputs should output text to stdout/stderr,
which will be forwarded to the serial port, and then shuwdown the
os after completing.

Now, ctest works and runs the tests directly in qemu. This aspects
works by having a wrapper script which calls qemu with graphical
mode disabled and the proper kernel command line, and then greps
for FAIL to see if the test failed.

This patch also fixes FileWatcher on iros, which was for some reason
implemented completely wrong. This testing mechanism already revealed
1 bug.