libc: fix 2 bugs in pthread_cancel
The first was using checking whether cancellation was enabled using
flass & 1. The fix inverted a check for being disabled.

The second issue was not checking cancelation immediately when calling
pthread_setcancelstate(PTHREAD_CANCEL_ENABLE). This is fixed by directly
calling pthread_testcancel().

Note that cancelation in libc is not spec compliant at all, as the
only cancelation point is pthread_testcancel(), and nothing else will
ever cancel a thread.