kernel: allow passing arbitrary arguments to the start program
Previously, the test= kernel command line was used to specify a
test should run. But this can be generalized by allowing any program
to be the start program. In real uses, this could even be the shell.
This also, lets arguments to be passed to the start program, using
the start_args= command line. The arguments are pipe separated, for
lack of a better parser. Currently all command line parsing is ad-hoc,
and fixing it is very painful because the kernel is written in C.