libc: optimize i686 system calls, and fix pthread bug
Now, system calls with less than 3 arguments are handled directly
using inline assembly. This is a nice performance win, and more
importantly makes the pthread code for exiting a detached thread
actually work on i686. Previously, this would crash because it
tried to access the stack after it was unmapped. This no longer
happens because the system call is properly inlined.