tests: introduce notion of a scheduler delay for use when sleeping
Lots of low level thread and kernel tests want to be able to spawn
threads or processes, let them for a bit, and then interact with
them. Ideally this would be solved using pipes or explicit
synchronization, but will work fine given a long enough iterval.
This patch makes this delay configurable, which would allow this
delay to be longer in CI then when running locally, and overall
results in making the tests cases run faster.