build: make iros tests have better ctest integration
Since we can't run the tests executable directly to ask it what
tests it has, we instead just grep the test source files for
TEST() macros. But this works almost as well.

However, on CI, this is disabled since doing so slows down the
tests signifcantly, as the overhead of starting qemu and booting
the kernel is much larger than the actual test cases themselves.