ports: add port of the ninja build system
This port is different in that it uses CMake, but this actually makes
things easier to work with now that we have toolchain files for the
system.

This port has one patch, which disables a call to the somewhat typical
getloadavg() API, since this is not yet implemented. In addition, the
effort to implement posix_spawn() is useful since ninja uses it.