libliim: implement Maybe<T&> as a wrapper around T*
This patch enables the use of optional references throughout the
code base. This is particularly useful when implementing containers
which perform falliable lookup. Previously, this pattern was implemented
using raw pointers, which it much less ergonomic. Returning a Maybe<T&>
allows using its monadic functions and also automatic panicking if an
empty Maybe is deferenced.

This patch also extends the Maybe API to allow construction using
the specical type None, which is useful to force an empty construction,
and implements a unwrap_or_else() function which converts a Maybe<T>
to an Result<T, E>, which is needed to use Result properly.