libliim: rewrite the internal Variant mechanism
The previous implementation was ugly, wrong, and relied on nearly
copy pasted code from libstdc++-v3 (for the visit()) function. This
new implementation works much better, and should even be constexpr
enabled. It now uses a recursive union to store its values, instead
of using reinterpret_cast's.

This patch removes the multi visit() functionality, as it is actually
completely useless and excessively complicated to implement. The
new visiation mechanims works by building a table of function pointers,
one for each variant type, and passes the visitor an in_place_index_t
with the current index. This enables calling code to know exactly
what type it is being called on, which is communicated to the recursive
union to implement operations like destruction and swapping.