port_manager: add configuration and patch step
This patch fleshes a lot of things out, the first of which is the
configuration object. This class contains the target architecture
as well as the system root and build directories.

The port class now contains its own paths which refer to the port's
own build and source directories, as well as the directory in which
the port.json is found. This is used by the patch step to locate the
specified patch files.

The context object now has a method to change the current working
directory, which will be very useful for individual build steps.
Instead of providing a RAII type for changing and restoring the
working directory, passing a closure which runs with said working
directory is used. This makes the side-effect usage clearer, and
clearly deliniates the scope in which the change happens.

The patch step works by running git patch on every specified patch
file in the port json file, in the port's source directory.

For now, the default file structure port builds will be something like
this:

|--> ${IROS_ROOT}/build_${TARGET_ARCH}/iros/ports/${PORT_NAME}/${PORT_VERSION}
|------> src
|------> build

The advantage of putting each port in its own directory is that cleaning a port's
build after the fact can be as easy as running rm -rf.