libcli: allow specifying defaulted flag values
Normally, optional flags have value types with LIIM::Option<> in them.
If the value instead has a proper default value, then the Option<> is
not needed. Note that the default value is specified in the parsed
argument structure rather than directly in the call, since this way is
just normal c++ programming. In addition, the user can specify a default
constructor which can do complex initialization.