meta: update remote container configuration and docs
Somehow, simply passing the DISPLAY enviornment variable into the
container is enough to get it to run qemu graphically. This is very
nice, since now means that no sophisticated work-arounds are needed
to run in a container. In addition, this patch publishes some vscode
configuration files which are very helpful for developing the os.