libliim: implement a new vector class
This class offers several improvements over the old LIIM::Vector class.
The first is using size_t for the size type instead of int, which was
simply terrible. The second is that this will work under constexpr
context. And the third is that it aims to implement a superset of
std::vector, but implement the same interface when reasonable. This
manifests itself in that the class has methods with the standard c++
names (push_back, insert, erase), instead of (add, remove, etc.).

The API does differ, the first being at() returns an Option<T&>
instead of throwing an exception on invalid accesses. This is quite
logical given exceptions are disabled anyway. In addition, insert and
erase methods allow passing either an iterator or and index into the
method call, which is more flexible while still implementing the same
functionality.

Currently, this implementation is missing methods to insert and construct
from an arbitrary iterator. It also does not optimize cases where the
underlying T is trivially destructible, or trivially copiable.

A strange property of this implementation is that the data() member is
non-constexpr. The reason for this is that the Vector internally stores
a MaybeUninit<T>* which it allocates directly. This allows not calling
the default constructors when allocating the backing store. However, casting
this storage to a T* is not allowed, and handling a pointer to the first
element makes g++ complain (as it probably should). This can only really
be fixed by using std::allocator, which is manually intercepted by constexpr
evaluator in g++, and allows allocating uninitialized storage. However, this
would make the vector type have to worry about allocators and no longer be
usable in libc (since libstdc++ can't be used). As such, its best to just
have data() not be constexpr. Any code using this type in constexpr context
can just use iterators and really has no reason to even use this function.