port_manager: allow build steps to be skipped
Certain steps, like downloading the program source, applying patches,
and configuring the port have side effects which make them not re-runnable.
So, we let them be skipped so that build will always work. This will be
very useful when ports have dependencies, since their builds will be
cached by this mechanism.

To know whether or not to apply patches, the port manager creates a
hidden file .did_*.patch every time a patch file is applied in the
port's downloaded source directory. This is pretty effective but is
not perfectly precise. It may be better to make every port that
requires patching just be a forked git repo, and then the patching
step would become useless.