port_manager: add a more sophisticated process spawning mechanism
The process object is kind of like a builder object, but instead of
having a build operation, it has a spawn_and_wait() operation. It can
be constructed either as a shell command (ran by sh -c 'command'), or
by taking a list of words explicitly, which and ran using evecvep.
This will in the future at the very least allow setting enviornment
variables, and may also handle input/output.

This uses posix_spawn internally, since this can be more efficent than
using fork(), and then waits on children using waitpid().