libliim: fix huge issue in Variant::swap()
The code attempted to store an r-value to an object which it immediately
destroyed. It then tried to use this r-value to construct a new object.
This patch fixes this issue by storing a pure value, which is moved into.
However, since Variants can hold references, explicit moves are replaced
with forward<T&&>(), since this only moves objects, and not references.