libliim: allow inserting collections directory into NewVector
This completes pretty much completes the set of features NewVector
needs to support in order to provide the vector api specified by
the c++ standard. In addition, iterator pairs do not have to be
passed, since the new container mechanism can replace this.

However, there is a showstopping bug in because of tempory variables
which prevents construction directory from collections from working
properly.