libliim: make invalid calls to range() return an empty range
In rust, doing something like:
```rust
for i in 4..2 {
    println!("{}", i);
}
```
will result in never executing the for loop. The same can be said
for python. However, the previous implementation would instead loop
from 4 to 1, by incrementing 4 until it overflowed back around. The
new method is a lot safer for developers to use.