libliim: implement tuple_map() and fix Tuple::get()
tuple_map() takes a tuple and applys a callback function on each
of the tuple's elements, returning a new tuple with the exact type
the return function gave.

In addition, Tuple::get() is fixed so that it works properly when
the tuple contains rvalue references. For some reason, the previous
version would attempt to return rvalue references when a prvalue was
needed (apparently). By duplicating the code for specific overload
scenarios (const&, &, &&, const &&), it makes it much easier to get
correct (especially since the problem makes little sense to begin
with).