libliim: implement zip()
This utilizes the tuple_map() function to provide arbitrary zipping
of any container. This essentially works by defining a ZipIterator
with simply stores a Tuple of actual iterators. Then, deferencing
is done by calling tuple_map() and applying operator* to each iterator.
Like in rust, if the iterators are not the same length, zip() will
simply stop early. Note that the Zip container is always unsized
even though it could be all of its component containers were sized.
The main reason this is not implemented is because it would require
excessive amounts of work, considering both specifying this constraint
and computing the minimum size after mapping a tuple both require
even more TMP tricks.