libliim: begin work on a new HashTable implementation
The internal implementation is essentially taken from google's
flat_hash_map. This aims to replace the old HashSet and HashMap
which was not closely related to the standard c++ API. In additon,
they didn't provide basic functionality like iterators (instead using
a lambda for_each method).

This API differs from the standard in that supports for containers,
find() returns an Option<T&>, pretty print formatting works, the
containers are constexpr, and a more uniform hashing API is provided.
It also aims to support heterogenous lookup.