libliim: implement transparent hash lookup and insertion
Transparent lookup allows you do query a HashSet<String> using a
StringView, ensuring the lookup will never allocate anything. In
addition, transparent inserts allow you to insert a StringView into
a HashSet<String>, and it will only allocate a new string if the key
was not present. This requires that the set type be constructible from
the type being inserted. However, since this is specified directly
with concepts, things should be clear to the user.

Transparent lookup is done by having the HashFunction for StringView
export a tuple of matches as follows:

```c++
using Matches = Tuple<String, const char*>;
```

Then the HashableLike concept will say that a StringView is hashable like
a string if and only if the StringView hash function declares it matches
with String. Note that lookup additionally enforces the constraint that
StringView and String be directly comparable.

To implement transparent insertion, a new primitive is added for
the Detail::Table implementation, called insert_with_factory(U&&), which
first performs a lookup using the passed argument, doing nothing if the
key is already there (but returning it the found reference), and if the
key is not already there, calling the passed in factory function with
the passed passed in key as well as the location in memory to call
construct_at() on. Making directly calling construct_at necessary is
to prevent any constructed objects from being moves/copied for no reason.

Furthermore, this insert_with_factory() function should be a strong
enough primitive to implement a proper HashMap on top of Detail::Table.