libliim: implement a much improved HashMap
This handles important aspects of the c++ unordered_map API, like
insert_or_assign and try_emplace, as well as providing transparent
lookup for all operations, and of integration with the containers
API.

This patch additionally had to add lots of extra logic to Pair in
order to properly support everything the map needs too. In particular,
peicewise construction.

Furthermore, the zip function now returns a pair when there are only
2 items being zipped together. This allows zip() to work seemlessly
with the constructor of the map.