libliim: rework new containers to use new construct helper functions
The big idea is that instead of having constructors for types like
Vector and HashMap, these classes will provide static create factory
functions, which in the future will return Results, and thus be able
to report memory allocation failures. However, for now, it is just
and alternative to a constructor.

This patch additionally introduces some new container methods, the
most important of which is insert(), which inserts another container
into some other container. This is used by a generic collect() function
which takes in any container and converts it into the passed template
parameter. Classes like NewVector provide a convience collect_vector()
function which deduces the NewVector template arguments for you.

Adding these overloads requires adding dummy methods to Hash::Map
and Hash::Set which take a const iterator as their first parameter
(this aligns with the c++ standard anyway, though). In addition,
library calls to emplace() now use the helper functions create/create_at
instead of directly calling the constructor, which ensures they work
properly with the new static factory functions.