libliim: use std::move and std::forward instead of our own
Note that for compatibility with usage in the kernel/libc, these
will be defined by liim inside the std namespace, even if the proper
c++ header it not available.

Doing this solves lots of headaches with ADL when attempting to call
move() on std types (like initializer_list<> and coroutine<>), which
is not solved.

Additionally, this allows static analysis and compiler diagnostics
to understand the move() == std::move() and forward() == std::forward().
This revealed 3 pointless uses of move() in the code base, just through
compiler warnigns.