libliim: rewrite Result<T, E>
This patch makes several improvements over the old implementation.
The first is that the Ok<T> pattern is removed, and instead results
are directly constructible from values. This is much less verbose,
and will allow generic code work better with results.

The second improvement is making Result<> move aware, which means most
of its methods are now rvalue qualified. This makes it more clear that
Results should be moved and not copied.

Additionally, the interface is now much more compatible with std::expected,
and the implemenation uses a union directly instead of using LIIM::Variant.

Finally, Result<void, E> is now valid, which makes a lot of code less verbose
and more clear.

This patch also begins work on falliable construction, particularlly by
implementing some concepts as well as a create<T> overload which takes
Results as input, and only succeeds if all the Results have values. This
is accomplished using the tuple apply() function, as well as
tuple_result_sequence(), which takes a tuple of results and produces
a result of a tuple. This functionality also collects error types into
a single error capable of representing all errors encountered, which
should usually is Error<>.