libliim: implement sort, insertion_sort, and is_sorted algorithms
The sort() implementation calls insertion_sort() for small ranges,
and otherwise uses quicksort. For comparison, these functions take
a three-way comparator, which differs from traditional c++, where
these functions take a less-than comparator. However, the three-way
comparator is more clear, and should be preferred for modern c++.

This patch additionally changes PriorityQueue to use a three-way
comparator, and changes it to default to a min-heap instead of
a max-heap. This is so that priority queue's are considered
sorted without needing to specifiy a comparator.

Also, new iterator adapter classes are added which make wrapping
an iterator or making a continuous iterator require much less
boilerplate.

The iterator wrapper also enables sorting transformed ranges, using
the new swap_contents() customization point. This allows for wrapper
iterators to forward to their parents, and will allow customer iterators
to have correct behavior when their underlying contents should be
swapped.