libliim: add a type erased string class
This type allows a uniform interface for accessing strings, which
may or may not be heap allocated. So a constant string literal will
be able to be erased, and so will a regular String. This is important
for the type erased LIIM::Error class, since error types may need
to materialize a string when queried for its message, while other
types, like SystemError, must not allocate, and will instead return
the string literal accessed via strerror().