libext: add a NewFile class to replace Ext::File
The old Ext::File does not have value sematics, by instead returning
itself in an UniquePtr. This was done do to the lack of a proper Result
type at the time (UniquePtr can be null). This new class makes a lot
more sense, and returns actual Results when performing fallible
operations.