libliim: begin implementing a PathView class
This class implements a logical PathView, which does not own its
underlying data, and does not interact with the file system. It
provides uniform iteration over components, while filtering out
extraneous '.' components. '..' components cannot be ignored because
symlinks exist. It additionally provides accessors for path data, like
parent_path(), filename(), extension(), is_relative(), etc.

This aims to replace the old Ext::Path class which required all paths
to be canonicalized using realpath(). This is needlessly inefficent,
and makes anyone using the class have to consider the case when creating
a path fails. In addition, Ext::Path stored a vector of components internally,
which meant that it was memory wasteful and required work to turn into a path
suitable for use with kernel file system APIs. Using this non-allocating lazy
alternative means that users can punt dealing with non-existent paths until
they actual use it.

In addition to this PathView class, a complementary Path class will be added
which owns its underlying data, and can be mutated.

These APIs are inspired by both the rust standard library and the c++ filesystem
API.

https://doc.rust-lang.org/stable/std/path/struct.Path.html#method.components
https://en.cppreference.com/w/cpp/filesystem/path