libunicode: add and use ability to determine width of a string
This patch adds a new unicode code generator, which parses the
unicode character database, and produces a data file which can be
used to lookup the value of a character property at run time.

In addition, there is a new library called libunicode which uses
the code generator to implement actual lookup functionality.

In this patch, only support for querying the EastAsianWidth property
as added, since this is very useful for creating TUI applications.

This patch also adds support for querying the terminal width of any
utf8 string or individual code point, by using the EastAsianWidth
property.

In addition, this capability is now used by libtinput to correctly
compute the display width of code points. In particular, this makes
the text editor correctly render the test_utf8_view.cpp file as well
as the newly added test_terminal_width.cpp file, both of which have
non-ascii characters in them, including emojis and east asian characters.
The text editor now correctly recognizes such characters as being
double-width, when previously they were treated as single-width.