libliim: add an Allocator concept and implementation
The Allocator concept differs from the one in standard C++ in that
the allocate() function returns both a data pointer and the size
actually allocated. This allows containers like Vector<> to get
optimal performance without needing to expose a realloc() function.
This is done by simply using the maximum capacity possible for a
given allocation.

For constexpr support, this patch also implements std::allocator,
which must be used internally in order to support constexpr allocation.
This patch also implements a standard allocator, which wraps std::allocator
to fit the liim Allocator concept.