libliim: add an initial implementation of Vector algorithms
The aim of this library is to create a generic interface for containers
with contiguous storage to use. Ideally, any new Vector-like type
needs only to implement the span() method, as well as reserve() and
capacity().

This should eventually replace NewVector's own implementations, and
also enable new Vector containers like a small buffer optimized
container as well as a static fixed size vector.

This also will be the basis for the new String's implementation.