libdi: begin implementing a replacement to liim
Since so much of the code already uses libliim, we get name conflicts
when trying to make new versions of classes such as vector and string.
To fix this, the only thing we can do is put the new functions into
a new namespace. This new library aims to cleanup liim to make the
code cleaner.