di: implement vocab::Tuple
This is a version of Tuple which is actually sized the same as a
struct with the same types. However, the current implementation
initializes the values in reverse order, improperly handles swapping,
and is missing many needed meta functions like make_from_tuple()
and apply().

However, it does provide a generic TupleLike interface, which other
types (Array<> and View<>) can use to implement structed bindings,
and interoperate with these Tuple<> meta functions.