libdi: implement Monads
Now, di::Optional is a Monad. This is done, for now, by inheriting
from di::function::monad::EnableMonad, and declaring suitable
and_then() and transform() functions.

The Monad unit functions works by using a CTAD constructor to the
template. The actual concept checking is done by the MonadInstance
concept, which asserts than both fmap and bind behavior reasonably
when passed the identity function.

A Monad is simply a template which can be constructed from Void, and
whose result satsifies MonadInstance.

For now, being a monad enables the use of ADL operators, with
>> representing `bind` and % representing `fmap`. For reference,
`>>=` was not chosen for `bind` because it is right-associative,
which prevents nice chaining. The current solution can also run
into problems, since % and << have different precedance levels,
but this can be avoided using (), named variables, or unit()
wrapping functions which could be done using `fmap`.