libdi: refactor ConstantVectorInterface to use free functions
This gets rid of the code duplication between the const and non-const
overloads in ConstantVectorInterface, and will enables other free
function vector algorithms to easily use this functionality.

In addition, this patch adds Optional<void>, which is a monad
with exactly 2 states (so a fancy bool). This facility is very
convenient when defining functions which return an optional.

For instance, consider the following function:

```c++
constexpr auto at(concepts::detail::ConstantVector auto& vector, size_t index) {
    return lift_bool(index < vector::size(vector)).transform([&] {
        return util::ref(vector::lookup(vector, index));
    });
}
```

By lifting the bounds check into Optional<void>, we can safely access into the
vector only when the bounds check is satisfied. If it is not, the returned optional
will be nullopt. This is especially useful in generic code, because it allows the
compiler to decude the return type of the function.

A "normal" version would look as follows:

```c++
template<concepts::detail::ConstantVector Vec>
constexpr Optional<meta::detail::VectorValue<Vec>&> at(Vec& vector, size_t index) {
    if (index < vector::size(vector)) {
        return nullopt;
    }
    return vector::lookup(vector, index);
}

template<concepts::detail::ConstantVector Vec>
constexpr Optional<meta::detail::VectorConstValue<Vec>&> at(Vec const& vector, size_t index) {
    if (index < vector::size(vector)) {
        return nullopt;
    }
    return vector::lookup(vector, index);
}
```

Notice this version requires duplicating code between the const
and non-const overloads, and requires using a meta-function to get
the required type to return.

However, it is clear that the second solution would be far more readable
than the first, and certainly should be used in for code which does not
require type-inference.