libdi: implement di::container::to<>(), and di::create<>()
The generic factory function di::create<T>() will either construct
an object of type T directly, call a customized create_in_place
definition, or call a types static ::create() factory function.

In addition, CTAD is simulated using the deduce_create customization
point, which takes an in-place template as its first parameter.
As such, the di::create<di::vector>(di::range(6)) is valid, and
the concrete type of di::vector is deduced.

di::container::to<>() works on top of the create mechanism, to enable
containers to be constructed from any arbitrary views in a pipeline.

In the future, the to<> function could have additional ways to
create containers, including manually inserting into them or
calling an iterator-pair constructor (but this is mostly useful
for interacting with std types).