libdi: properly support fallible allocations in Vector algorithms
This involes using the helper function as_fallible and try_infallible.
In addition, invoke_as_fallible can be used to properly handle the
case when a function returns void or Expected<void, E>. Finally,
the container::sequence() is provided to implement monadic sequencing,
which only works for Expected<void, E>. That algorithm could be made
more general, but for now it is unimportant.