iris: begin work on a new kernel written in C++
Unfortunately, the previous kernel written in C is very difficult
to work with, mostly due to C's lack of abstraction and my own
inability to compensate for it.

The idea is to write a new version, the Iris kernel, which uses C++
for RAII and coroutines, which should enable the code to be written at a much higher level
or abstraction.

For simplicity when starting out, the limine boot loader is used. This
allows us to write a higher-half x86_64 kernel without writing a single
line of assembly. This is very nice, since we won't have to worry about
setting up processors until SMP support is added.

This commit is simply a hello-world kernel, which also sets up the build
system and tools to be able to build and run kernel images easily.

This commit additionally relicenses the code from GPL 3.0 to the BSD
2-Clause license. This should be legal, since I am the only contributor
to the code.