libdi: implement di::BitSet and di::BitStruct
The di::BitSet API mirrors that of std::bitset, and di::BitStruct
provides an API for defining bit-packed structures without using
C++ bit fields.

This is an improvement because C++ bit fields have several problems,
with the most obvious one being the fact that their ordering is
implementation defined (and in practice is reversed on bit endian
architectures). In addition, grouping the elements of a bit field
together makes it more clear that they are not regular struct members.