libdi: cache important values in RBTree
This includes the minimum and maximum nodes, as well as the size.
This makes operations like front(), back(), begin(), end(), and size()
constant time. As consequence, the TreeSet<> data structure can be used
to implement a PriorityQueue, although it would not be the most
efficent implementation.

Note that standard C++ std::set does not support front() and back(),
which is especially unfortunate considering the container is sorted.