libdi: prevent reconstructing dangling containers
Because the reconstructed container for types like di::Array and
di::Vector is a di::Span, there were dangling references when using
expressions like:

```c++
auto x = di::Array { 1, 2, 3, 4, 5 } | di::take(3);
```

This is fixed by preventing reconstructing containers which are not
borrowed in di::take() and di::drop(). However, this check is only
required because the input container is dropped in those functions.

For an expression like di::split(), which will return its own view
type which holds a reference to the container it is splitting on,
this check can thankfully be avoided.