libdi: implement di::split()
As shown in the test cases, this allows splitting by value or sequence
of values. It also makes use of the reconstruction customization
point, so that splitting strings yields string views, and splitting
vectors yields spans.

This is inherently dangerous without a borrow checker, since the return
values of split will always reference the container being split on. Since
SplitView is not a borrowed container, code which uses split locally in
a chain of pipes will never reference invalid data.

However, users can easily cause problems once they pipe into di::to<di::Vector>
on the result of a split, since this will destroy the underlying split view,
which may be keeping ownership of the String data being split on. In this case,
there would be a Vector of dangling StringViews.