libdi: implement LinkedList and Queue container adapter
Unfortunately, the LinkedList appears to be unusable in constexpr
contexts. Whenever a function attempts to return a LinkedList,
GCC complains with the following error message:

```
error: modification of ''result_decl' not supported by
dump_expr<expression error>' is not a constant expression
```

Since the code compiles and executes correctly when not in
constexpr contexts (and with asan and ubsan enabled), this is
probably a bug with GCC constexpr evaluator.