libdi: implement di::slide()
This deviates from standard c++ by using container reconstruction
to derive the value type. In practice, this means the sliding windows
of a String type will be StringViews, and the sliding windows of an
vector type will by Spans. This is pretty nice, although it requires
a slightly more involved implementation.