libdi: implement di::TreeMap and MapInterface
The TreeMap class is simply a normal RBTree with a specialized
comparator which compares only elements keys. Additionally, the
value type does not have a const Key member so that the value
type is movable. In the future, it may be wise to adjust the
container's iterator to not return mutable key references, but
internally the keys definitely should be mutable to prevent
useless copies.

Although the MapInterface is staggeringly large, it is only because there
are a huge number of overloads required by the standard c++ map interface.
In the future, when implementing a HashMap, only a handful of functions
will be need, so the large code duplication in SetInterface and MapInterface
will be justified.

It also implements heterogenous lookup, which allows searching for
StringView's in a TreeMap<String>.