aoc: add a small number of advent of code solutions
These problems serve as simple test problems which exercise the
di library in nontrivial ways. When writing these solutions, I've
fixed a large number of issues with the library. Additionally, this
has driven development of a basic userland library for CLI parsing
and output logging.

The biggest issues encoutered so far are the lack of source location
being printed by the ASSERT() macro on failure, lack of error messages
in general, sketchy format implementation, and horrendously long error
messages while using views.

Note that for the last one, actual incorrect client code is rare, and
when errors do occur, they are usually shorter.

In the future, since these problems can be used for regression tests,
since they have a known output but involve nontrivial use of the library.