libdi: implement some basic parts of p2300r5
This commit just implements some concepts and customization points
defined in that paper, but does not implement any algorithms.

The main concern for this implementation is whether or not this model
make sense without using c++ exceptions (or more importantly, is this
usable for a async first OS kernel).

So far, most customization points require their implementation to be
noexcept (set_value, set_error, set_stopped, start). However, both
connect and schedule are fallible. The most likely adaption for to
disable exceptions is to allow these CPOs to return di::Result<>, and
treat the error channel as having a di::Error (type erased status code)
instead of std::exception_ptr. However, this may prove problematic
for connect() because doing so requires movable OperationStates's, which
are likely to contain synchronization objects (std::atomic) which are
not movable.