libdi: make formatting fallible
This is necessary because certain format operations can fail. In
particular, formatting with a run-time width, as well as formatting
integers as characters. And when passing a runtime format string
as well, such as allowing the user to specify a time format.

It is possible to make a certain subset of format operations
infallible, but this requires an excessive amount of work,
especially when considering that formatting requires either
allocation or IO, both of which should be considered fallible
in some contexts.

However, debug logging is infallible, meaning that an errors
that occur will be ignored. Since the format string validation
mostly occurs at compile-time, the convience should be worth it.