libdi: implement co_await support for sender in di::Lazy
To be able to support co_await'ing senders, a sender 3 channels
must be mapped into coroutines. Since exceptions are disallowed,
the error and stopped channels are fully out-of-band, and call
specific functions on a promise type directly.

In the default implementation, this simply unwinds the coroutine
stack, up to the caller. For example, when calling di::sync_wait()
on some coroutine task, if a unhandled error occurs at any point
within the coroutine's execution, the sync_wait() will complete
with a Result<> that has the error which occurred. This happens
majically anytime co_await is used, which means that errors
automatically propagate like exceptions. In order to actually
handle errors inside the coroutine itself, a sender adapter needs
to be used. Note that this is the default behavior, but it is
most likely possible to have co_await's of sender's return Result<>
directly.

Under this model, it is impossible to tell whether any di::Lazy
is infallible, and so essentially every di::Lazy can now complete
with a error.